Integer to Long

+++ src/main/java/Main.java
@@ -39,7 +39,7 @@ public class Main {
       return bencodedString.substring(firstColonIndex+1, firstColonIndex+1+length);
     } else if(bencodedString.charAt(0) == 'i'){
       int len = bencodedString.length();
-      return Integer.parseInt(bencodedString.substring(1, len-1));
+      return Long.parseLong(bencodedString.substring(1, len-1));
     } else {
       throw new RuntimeException("Only strings are supported at the moment");
     }
